Servicehandler Tests
this commit includes :
- Add rpc-state-modify on servicehandler.yang to test
rpc-service-restoration which can only be launch on service with
operational state 'outOfService'.
- Move stubpce.yang and stubrenderer.yang to stubmodels maven project
under tests folder.
- Create specific profile for servicehandler tests in tox.ini file.
So now to only launch servihandler tests you have to use this
command : 'tox -e servicehandler' in tests folder.
- Use parameterized loggins.
- Add 5 more tests to 'tests_servicehandler' python file.
- Mofify some errors on service-create rpc implementaion.
- Add python file 'tests_servichandler.py' to test servicehandler
functions.
- Modifiy stubpce and stubrenderer yang model to add RPCs from
servicepath version 1.5. This modification came from having two version
of servicepath (1.5 and 1.6) in the model directory does not allow me
to access some RPCs. For example trying to get service path list for
1.5 version failed caused the url is the same :
http://localhost:8181/restconf/operational/transportpce-servicepath:service-path-list
So REST API seems to use the last version which is 1.6.

Change-Id: Ief8aab19750fe2959407b90772df8a8a6280e73d
Signed-off-by: Martial COULIBALY <martial.coulibaly@gfi.fr>
Signed-off-by: Olivier RENAIS <olivier.renais@orange.com>