Fix some bugs in topology tests
Remove in some tests the manual node deletions in openroadm-network.
This action is already performed by TransportPCE when disconnecting the
topology-netconf device. Related tests should no longer report an error.

NB: test_05 in test_topoPortMapping.py still fails because of a bug in
the creation of xpdr device in the openroadm-topology.

JIRA: TRNSPRTPCE-86
Change-Id: Icf0eb7b38a2bd6f82b24a64ed32c44238fdb9087
Signed-off-by: Gilles Thouenon <gilles.thouenon@orange.com>
Co-authored-by: Christophe Betoule <christophe.betoule@orange.com>