Update patch set 1
Patch Set 1: Code-Review+1

(2 comments)

Thanks for the good job!
Looks fine to me from a functional point of view
Might be worth to add comments on the use of the otn-service-interface to store history of interfaces and connections created using service-name as a key.
Just as a reminder for future enhancements (not in this sprint): 
_Add later in portMapping.updatemapping trib-port and trib-slots associated with ODU4 supporting interface for topo update purpose
_Activate later in OpenRoadmInterface221 in the interfaceBuilder .setCircuitId for alarm correlation
_Later on, complements createOpenRoadmOdu2interface to handle in addition to 10GE, OTU2 supporting interface.

Patch-set: 1
Reviewer: Gerrit User 7217 <7217@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1