HoneyNode Java 11 support for 121 devices
- honeycomb 1.19.04 upgrade
- rename 2.1 folder into 1.2.1 to reflect the openroadm device version
  rather than the openroadm common version
- update resconf honeynode config to allow remote access
- samples config files update
- delete old testtool 1.2.1 sample configs
- added support for openconfig models
- remove JDK8 gate support from tox and associated scripts

NB: script installJDKCentos.sh is just a workaround to have the right
JDK. I think that configuring the version of the JDK in Jenkins job
should be a better solution. By default in
transportpce-tox-verify-magnesium job, JDK is set to System which seems
not to be JDK11. In dropdown list, opendjdk 11 is available.

JIRA: TRNSPRTPCE-192
Signed-off-by: Emmanuelle Delfour <emmanuelle.delfour@gmail.com>
Change-Id: Ia30ceb05c927937d9b4fc81d42d91b7558cd1dfd