Make PCE optical node validation fail when no PPs are available
Currently path computation succeeds even if all external PP ports on an end SRG are already used. Then the renderer will fail during service setup. It seems to me that it would
be better to have the path computation fail. This is a small proposed change to make that happen.

Change-Id: I92308cc7f12f9957f376c4d10b3d92323e865481