Update patch set 9
Patch Set 9: -Code-Review

> Patch Set 9:
> 
> ready to merge as far as I am concerned. Waiting for someoneelse's confirmation.
> Caution: OLE PCE Junit on going development impact this change.

Gilles'feedback is that there is a bug at the kernel project level with thos new models that makes mandatory the ethernet subrate parameters with OC services. It is mandatory for eth services but it doesn't exist with OC services (or something like that).
This needs to be clarified with a JIRA bug and commented in E2E tests and in the commit message before merging.

Patch-set: 9
Reviewer: Gerrit User 6731 <6731@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=0