Read all LLDP neighbours even if one fails
Currently in the readLLDP function, when looping over all neighbours in
nbrList, if createR2RLink for one neighbour fails, the whole function
returns which means that the rest of the neighbors in nbrList are never
processed. I think it makes more sense to continue the loop and create
R2R links for other neighbors. This patch proposes to do that while
still returning false if one createR2RLink fails.

Signed-off-by: Jonas Mårtensson <jonas.martensson@ri.se>
Change-Id: I73d587cc52c1b11e449baf3c69c9179a7dc2483d