Avoid converting spanloss to double when calculating target power value
Currently spanloss is measured and set without any fractional digits.
We would like to have one fractional digit but this creates problems
because of the conversion to double when calculating target power
resulting in the power value having more than the two fractional digits
allowed by the yang module. This patch proposes to skip the conversion
to double and only use BigDecimal to avoid the problem.

The proposal to measure spanloss with one fractional digit will be
submitted in a separate patch.

Signed-off-by: Jonas Mårtensson <jonas.martensson@ri.se>
Change-Id: Ibd13c991d43f2e94739710412e8edadcc0c537bf