Update patch set 3
Patch Set 3:

> Patch Set 2:
> 
> (3 comments)
> 
> apparently something prevents the feature from being installed in the gate.

This is not related to the gate. That's what at said during the last standup : sometimes after installing the feature, the restconf bundle does not restart correctly => that's the reason why we have 404 http error. I also said that we have to investigate on that and it will be done in another change. As it's occured randomly, it will take more time than usual ;-)

Patch-set: 3