Update patch set 3
Patch Set 3:

> Patch Set 2:
> 
> (3 comments)
> 
> apparently something prevents the feature from being installed in the gate.

It's not related to the gate. Sometimes after installing the feature, the restconf server does not "restart" correctly. that's the reason why you have 404 http errors. I talked about that during the lats standup. I said that we need to investigate. It will be done in another change. As it occurs randomly, it will take more time than usual to find the reason.

Patch-set: 3