Update patch set 3
Patch Set 3:

(2 comments)

Thanks Bala. See my comments.
According to the yang model, circuit-pack-name and port-name in the list of xponder should be mandatory. However, I see that with honeynode, it is possible to start the node even if that leaves are missing (issue with mdsal/yangtools? managing the "refine" keyword?).
With a full compliant device with OpenROADM spec, this case should not appear. But, agree with you to put more protection in the code.

Patch-set: 3
Reviewer: Gerrit User 7446 <7446@7fc14799-209e-464c-9743-7a06c2c21a81>