Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Should be more relevant on stable/magnesium branch than master, all the more than master is being migrated to Aluminium.
> 
> Until flexgrid code is still based on Mg, I don' think so.
> Otherwise, without this change, we could not fix the current issues on it; for instance some Junit and RSPN tests in failure

I agree with Gilles comment. This change dedicated to Mg development is more relevant to be pushed on stable/magnesium.
This brings up a more general question of how to push the changes on the master branch and the other branches.
Wouldn't it be more practical to push the code to the corresponding branches and cherry-pick by the same time the change on Master branch if it impacts future developments?
This is in order to avoid having to cherry-pick a large number of changes during the release process and encounter possible performance issues with the gate.

Patch-set: 1
CC: Gerrit User 7085 <7085@7fc14799-209e-464c-9743-7a06c2c21a81>