Add OpenROADM 7.1.0 device and common models
- add OpenROADM device and common models version 7.1.0
  (previous models 1.2.1 and 2.2.1 are kept)
- keep only essential OpenConfig models used by OpenROADM
- add a NOTICE and an openconfig.LICENSE files in the jar ressources

OpenConfig YANG models are published under the Apache License version 2
a.k.a. ALv2 but they do not precise it in their description.
Though Alv2 allows redistribution under other licenses, it includes a
redistribution clause (4) that instructs to give any other recipients
of the work or the derivative works a copy of the original license.

JIRA: TRNSPRTPCE-225
Change-Id: Idebe4e20f0cb0554dd572c4c667d692c8a6dfcb5
Signed-off-by: guillaume.lambert <guillaume.lambert@orange.com>