Unplug savagely a GNPy link to the PCE module
MDSAL codecRegistry Serializer implementation can no more be used
directly as it is done in the PCE for GNPy. This operation must be done
now through the BindingDOMCodec OSGi Services instead.
However, this adaptation is easier to proceed after the PCE migration.
Let's unplug this part of the code for the moment.

JIRA: TRNSPRTPCE-292 TRNSPRTPCE-302 TRNSPRTPCE-314
Signed-off-by: guillaume.lambert <guillaume.lambert@orange.com>
Change-Id: If841413637b1d6983bdae49d835f6ccb574036d8