Update patch set 2
Patch Set 2:

(4 comments)

Hello,

Thanks for your comments. I fixed the "revision should stay sticked" but for the other I'm sure to understand. The goal of the change is to fix sonar issue and not making code optimization. I remember change I pushed where you said that there was "à boire et à manger" pointing the fact that the change contained to many type of improvement. 
So, for the current change, I follow your advice and did not code optimization. For me, following your rules and recommendation, it should be in a dedicated change.

Emmanuelle

Patch-set: 2