Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> For me, this change is unnecessary on master. We already have change 93026 which proposes a migration of master towards Silicon. It would be preferable to solve disagreements on that change and merge it rapidely instead of keeping master sticked to Aluminium. Moreover, there is no broken dependency which imposes to have now this transition step...

Noted but I disagree. Here are my explanations.
There are still a couple of issues that are not addressed in change 93026 and I have no visibility on when they will be fixed so that we can merge it.
Since yesterday, we have a karaf runtime issues because our Al *-SNAPSHOT upstreams deps were rebased on these new Al SR1 dev core dependencies.
So until either this change or 93026 is not merged, the gate remains unusable...
We have still in master ongoing works based on Al SR1 dependencies and they need this update to use the gate.
This is the quickest option to make the jenkins releng gate works back again.

Patch-set: 1
CC: Gerrit User 7231 <7231@7fc14799-209e-464c-9743-7a06c2c21a81>
CC: Gerrit User 2850 <2850@7fc14799-209e-464c-9743-7a06c2c21a81>
Reviewer: Gerrit User 6731 <6731@7fc14799-209e-464c-9743-7a06c2c21a81>