Update patch set 5
Patch Set 5:

> Patch Set 5: Code-Review-1
> 
> (13 comments)
> 
> Hello,
> 
> Please find my review. 
> CR-1 is due to checkstyle violation + use of implementation instead of interface in method signature.
> Regards

Correct this is reported by mvn-verify tox job too.
I only tested locally that PCE and gnpy functests passed.
At this step, I was more interested in getting the whole functests result from the gate here and some feedbacks on 92812.
Though, when making this change public, I forgot to mention that by putting this change in WIP (as undertoned in 92812).
Note there is also a functest failure in otn E2E that would justifiy a vote -1.

11:53:26 py3 run-test: commands[22] | nosetests --with-xunit transportpce_tests/2.2.1/test_otn_end2end.py
11:56:05 .........................................F.......F....F..........

Anyway, thx for the feedback, I will try to use the intfce rather than an impl here and see if it improves the situation but I suspect a deeper problem underneath.

Patch-set: 5
Work-in-progress: false