Update patch set 13
Patch Set 13: Code-Review+1

Could we please merge this change. I have no doubt that if we want to find something to block it, we can find it. However I don’t see some good reasons to do so. I see that Emmanuelle, spliting her work into different patches tried to do all she could to simplify the review process, thus adressing the Ptl’s major concern. This is a M patch, to which comment to Giant patches do not apply. Most of the very long comments I see are based on the assumption that we could backport some of the modifications made to other branches. However I don’t see here fundamental reasons that justify why we should do so. Maybe some modifications on test could be backported to other branches, but it is not strictly needed, does not add value and generate additional work. Blocking this change is counterproductive. Should I remind that the priority defined for this sprint is to prepare for Flexgrid and higher rates and pursue the development needed for MDONS use cases? Thanks

Patch-set: 13
Reviewer: Gerrit User 7217 <7217@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1