Update patch set 13
Patch Set 13: Code-Review-1

> Patch Set 13: Code-Review+1
> 
> Could we please merge this change. I have no doubt that if we want to find something to block it, we can find it. However I don’t see some good reasons to do so. I see that Emmanuelle, spliting her work into different patches tried to do all she could to simplify the review process, thus adressing the Ptl’s major concern. This is a M patch, to which comment to Giant patches do not apply. Most of the very long comments I see are based on the assumption that we could backport some of the modifications made to other branches. However I don’t see here fundamental reasons that justify why we should do so. Maybe some modifications on test could be backported to other branches, but it is not strictly needed, does not add value and generate additional work. Blocking this change is counterproductive. Should I remind that the priority defined for this sprint is to prepare for Flexgrid and higher rates and pursue the development needed for MDONS use cases? Thanks

Sorry Olivier, but what must be merged or not be merged is the exclusive matter of committers, and not yours. As well as the maintenance of the other stable branches.
I am very aware that this change works out-of-the-box on the master branch and brings a lot of very nice improvements.
But except if we decide to drop the maintenace of stable branches, I can't reasonably merge it.
Moreover if we plan to include subsequent pieces of code to the stable/aluminium branch in order to support flexGrid and B100G as we agreed before.

The best I can do to speed up the integration of this change is to give quick reviews on it. And I have done it. It's now the responsability of its owner to take them into account. I didn't have the feeling these modifications were such a big deal. But I might be wrong and, again, if the owner has difficulties to integrate them, I will be happy to help. I can also take care of doing it on behalf of the owner, but this is normally in last resort. After a quick talk with the dev team, we agreed to proceed this way this morning.

At least one of the improvement here, the removal of odl-mdsal-broker useless dependencies, is not related at all to the Silicon bump nor the tests and it must be backported to the other branch(es)
see https://git.opendaylight.org/gerrit/c/transportpce/+/93551
I agree the others improvements are somehow related to the Silicon migration but it is still highly wishable to backport them on the stable/branch too.
To do so, we need to keep a trace of the cherry-picks done by keeping the same change-ID. This can not be done if these improvements remains inside this change...

Patch-set: 13
Reviewer: Gerrit User 6731 <6731@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1