Update patch set 4
Patch Set 4: Code-Review-1

I don't see the interest of such a change. If lighty artefact for silicon is not available yet, we have just to use karaf as it was the case until few days...
And when lighty artefact will be available, you will just have to activate it again in tox if you really care.
I guess we are not in a situation where we have no other choice that doing things that looks like hacks.

Patch-set: 4
Reviewer: Gerrit User 7446 <7446@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1