Update patch set 4
Patch Set 4:

> Patch Set 4: Code-Review-1
> 
> I don't see the interest of such a change. If lighty artefact for silicon is not available yet, we have just to use karaf as it was the case until few days...
> And when lighty artefact will be available, you will just have to activate it again in tox if you really care.
> I guess we are not in a situation where we have no other choice that doing things that looks like hacks.

I understand you came to the same conclusion than me. The runtime problem is due to a problem in the lighty-core SNAPSHOT and not in the tpce lighty controller code here.
Note that:
- I already self-assigned a verify -1 and that in this kind of situation, a vote -1 is normally pointless.
- I am fine too with your proposal of using karaf till it works. cf TODO. The lighty enforcement here is more a convenient way to show test results with lighty activated without overloading the gate. The gate is particularly loaded as you noted in https://git.opendaylight.org/gerrit/c/transportpce/+/93564/ -> 3 timeout in a row on jenkins/releng...

About the interest of this change, it seems I need to detail it more in the commit message or the comments. Here it is.
This change addresses an omission in the Silicon migration.
The previous change haven't modified the lighty-app-parent dependency version in lighty/pom.xml. So the lighty controller code here was not even building since the version 13.0.0 left there is not usable with Silicon.
As you noticed, the version 14.0.0 is not available yet and the only remaining solution to make it build, is to  use the snapshot version (14.0.0-SNAPSHOT). But to make it usable, we need to build lighty-core master locally first because Pantheon does not propose it on maven central or any public repo that I know.
I agree there is still a problem at runtime and there is no question to do hacks here. We used to  work this way with lighty already in the past with Neon.

Without that change, it was impossible to identify the runtime problem here. 
Maybe I was a little too optimitisc but nobody is in a lee of a good surprise.
Anyway, this is still an improvement compared to the previous situation.
And if we need to have a ligthy version working at runtime, we can still:
- propose a fix to the problem to the lighty official repo
- wait that it is fixed by Pantheon during the Silicon release cycle

Does it make more sense ?

Patch-set: 4
Reviewer: Gerrit User 6731 <6731@7fc14799-209e-464c-9743-7a06c2c21a81>