Flexgrid for pce and network model
Flexgrid implementation in pce and network model.
- add ServiceRateConstant class in common module to centralize service
rate Uint32 format.
- update GridConstant with maps (service type and nb slots, service rate
and modulation format, service rate-modulation format and frequency
width) and constant for irrelevant wavelength number.
- add methods in GridUtils.
- rename methods in NetworkModelWaveLengthService to be compliant with
frequencies.
- disable NetworkModelWaveLengthService unit tests.
- move some methods from NetworkModelWaveLengthServiceImpl to
OpenRoadmTopology util class in network model.
- update pce with flexgrid spectrum assignment computation.
- update pce unit tests.
- add getBitSetData method to PceNode and remove checkWL method.
- update PceNode implementations (PceOtnNode and PceOpticalNode).
- rename methods in pce to be compliant with frequencies.
- update some log messages to be coherent with frequencies.
- add minFrequency and maxFrequency to PceResult.
- update functional tests.
- update topology test files set freq map value to have the smallest
lambda on the higher index of the freqmap.

About "disable NetworkModelWaveLengthService unit tests":
some updates are needed in test-common: path-description is a
grouping and is used only in notification. DataObjectConverter converts
only data so to be able to Deserialize path-description from json file
in unit tests, a solution could be to add a dummy yang model in
test-common with container using path-description.

JIRA: TRNSPRTPCE-230
Signed-off-by: manuedelf <emmanuelle.delfour@orange.com>
Change-Id: Ie22761e535e0669f614ac18ea801b5dac75dd362