Update patch set 18
Patch Set 18: Code-Review+1

> Patch Set 17:
> 
> > Patch Set 17:
> > 
> > (1 comment)
> > 
> > > Patch Set 17:
> > > 
> > > Hello,
> > > 
> > > There is too much "break" and "continue" in for loop. It does not improves the readibility of code. 
> > > 
> > > Regards
> > 
> > Hello
> > 
> > Yes there are many more break + continue.
> > Can you elaborate more on where and on which criteria it decreases readability ?
> > 
> > For example in the  version 121 file line 545-571 (base) vs line 555-586, the new structures looks much more readable to me because the nested blocks/strucutures depth is reduced from 4 to 1, what really eases the indentation.
> 
> Break and continue increase complexity of source code because you cannot linearly read the flow of statements. And it's even worse when they are in a for loop.

Please can you be more specific?

There is not one sole definition or interpretation of what is complexity.

According to sonarsource/sonarcloud/sonarlint/sonarQube documentation, complexity is a metric that can be measured in Java by counting the following keywords: if, for, while, case, catch, throw, &&, ||, ?.
See https://docs.sonarqube.org/latest/user-guide/metric-definitions/
In this meaning, the global adaptations and optimizations proposed here factually decrease the complexity metric of the code since they remove a few doppelgangers conditions and checks. 
( Though, when applied locally, this assertion has to be taken for a same functional perimeter since the contribution also adds a few (forgotten?) protections that should not be considered. )

Also if we take the default sonar rules in the related category ("brain overload") at
https://rules.sonarsource.com/java/tag/brain-overload/RSPEC-135 ,
it appears that this contribution:
- can effectively triggers the rule 'Loops should not contain more than a single "break" or "continue" statement' which is tagged as a minor issue
- but fixes a few issues related to the rule 'Control flow statements "if", "for", "while", "switch" and "try" should not be nested too deeply' which is tagged as critical
- and fixes a few issues related to the rule 'Expressions should not be too complex' which is tagged as critical too

When reading the details in the description of the first rule mentionned, it is stated that "continue" and "break" can facilitate optimal coding, what is coherent with the title of the change.
As detailed in JIRA, this contribution is a preliminary step that aims at aligning and optimizing the 3 current portmapping implementations. This step will ease a deeper refactoring to come that can address more easily the break/continue issue by introducing new functions as you proposed in a previous change.

Patch-set: 18
Reviewer: Gerrit User 6731 <6731@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1