Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> This change is no longer relevant here on master and on stable/aluminium since package of changes related to flexgrid has totaly rewritten this class.
> For me this change could now only be proposed for magnesium SR4...

@Emmanuelle @Gilles

I agree this has been completely rewritten for flexGrid.
But there are lots of considerations to object to this remark.

- This change addresses some remarks I made on network model 7.1 a fortnight ago.
And I clearly stated at that moment that I wanted to address them.

- I am not against pushing that for Mg SR4 but it has to pass by the master branch before anyway.
So we need to stack it with the current adaptations on FlexGrid.

- Also how was I supposed to know there were potential merge conflicts with Emmanuelle's change ?
Emmanuelle's work was proposed more or less simultaneously on a different branch (stable/aluminium), which is not a dev branch. I was added in CC only a few hours after proposing mine...
Can we decice to all work on master to avoid that kind of situations starting from now ?

Patch-set: 2
CC: Gerrit User 2850 <2850@7fc14799-209e-464c-9743-7a06c2c21a81>
CC: Gerrit User 7231 <7231@7fc14799-209e-464c-9743-7a06c2c21a81>
Reviewer: Gerrit User 6731 <6731@7fc14799-209e-464c-9743-7a06c2c21a81>