Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > hello,
> > 
> > I'm not sure that we need this in stable/aluminium?
> > Gilles, what is your opinion ?
> 
> Yes Indeed, I'm not sure this change had been targetted for aluminium. It goes more for higher rate (Silicon), and especially portmapping, renderer and olm 7.1 in order to perform first 7.1 functional tests.


It's effectively not clear whether 7.1 sims are needed or not for Al by itself, but I think this will be likely helpful to prepare OFC demo devs based on this branch.
Anyway, this does not change many things since there is no 7.1 functional test for the moment and the download of the 7.1 sim is inhibited by the next change.

In this situation, version tags on the associated JIRA item would have been welcome to state it clearly  but unfortunately we cannot create release versions in JIRA items since we do not have the admin rights on it...

Patch-set: 1