Update patch set 2
Patch Set 2:

(1 comment)

Thanks so much for this great work! Your expertise on high rate OTN interfaces is really appreciated. Looks good from my side (from what I can evaluate and my limited knowledge of OTUcn hierarchy). This a huge step towards OFC demonstration. 
I have one general comment about the naming of the interfaces. As R7.1 is also supposed to include any kind of equipment, including OTN switchs of lower granularity, I would suggest that the name of the interfaces clearly reflects the rate we are considering (example : rather than Eth, Eth400G).  Because at a time we will need to include all other interfaces at lower rates that will also need to be supported on other 7.1 devices. For OTUCn we could imagine passing n value as a parameter, from which will depend further processing. In the same way, the number of trib-slots may depend on the rate, as well as foic type. We don't need to consider all intermediate rates (<400G) right now, but we shall try to make everything so that further extensions do not imply to change the code for the interfaces that are already handled.

Patch-set: 2
CC: Gerrit User 7217 <7217@7fc14799-209e-464c-9743-7a06c2c21a81>