Update patch set 9
Patch Set 9:

Hi Javier,
Tapi functional tests are failing since you are introducing a totally different implementation of the abstracted tapi-topologies performed until now. I believed that you were implementing a new topology that would let the existing topologies as they were. But, actually this is not the case.
Existing tapi functional test suite could stay as it is if your modifications may be compatible with the existing. But, if it is not possible, you will not have other choice that redo enterely the tapi functional test suite to be compatible with your new implementation. But may be before that, we should validate your new tapi topology proposal during our next TransportPCE meeting?

Patch-set: 9
CC: Gerrit User 7446 <7446@7fc14799-209e-464c-9743-7a06c2c21a81>
Work-in-progress: false