Update patch set 8
Patch Set 8:

(1 comment)

I've taken care of rebasing this at the bottom of the stack since
I consider it is a thorn on the side we are carrying for too long.

For the moment, I've only put the first initial change of the stack (95965) on top of it for 2 reasons:
- fixing its commit message by the way
- see if it passes the gate before going further

This way, we should have more freedom to merge intermediate steps.

Patch-set: 8
Reviewer: Gerrit User 6731 <6731@7fc14799-209e-464c-9743-7a06c2c21a81>