Remove odl-netconf-connector-all dep. from feat.
odl-netconf-connector-all dependency in features introduces the other
feature odl-netconf-callhome-ssh that comes with the bundle
callhome-provider. This bundle is not needed by TransportPCE and
consumes a listening port. Though, odl-netconf-callhome-ssh also brings
a third feature called odl-netconf-topology, which is needed by the
OpenDaylight netconf driver to work properly with Karaf.

As a consequence, the dependency odl-netconf-connector-all can
advantageously be replaced by the two dependencies odl-netconf-connector
and odl-netconf-topology.

JIRA: TRNSPRTPCE-482
Signed-off-by: Guillaume Lambert <guillaume.lambert@orange.com>
Change-Id: I90b96f8f92bd6936aa028df4edf5a9171b2f7e6f