Update patch set 6
Patch Set 6:

Si SR2 will be released soon. This change must preferably be put at the bottom of the stack directly on top of the master branch so that it can be merged simultaneously with the one in the Silicon branch.

Patch-set: 6
Reviewer: Gerrit User 6731 <6731@7fc14799-209e-464c-9743-7a06c2c21a81>