Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

> Patch Set 1: Code-Review-1
> 
> Not in favor to disaligne openroadm models of the project from official ones. Would not it be possible to adopt another work around to wait for the next model upgrade (8 or 10)?
> Would it be possible on openROADM side to create a revision release (7.2) to bring this kind of minor corrections?
> I would much prefer corrections on official models comes from openROADM rather than having to tinker with the project models.

I agree with Gilles that this should better be addressed by OpenROADM in a minor release.
If we deviate from official models for any reason it should be stated in the models and the commit message.

JIRA ticket is not enough though it explains the reason.
Also it is not clear about who made the decision to deviate from official models.

Patch-set: 1
Reviewer: Gerrit User 6731 <6731@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1