Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> (3 comments)
> 
> > Patch Set 2: Code-Review-1
> > 
> > Not in favor to disaligne openroadm models of the project from official ones. Would not it be possible to adopt another work around to wait for the next model upgrade (8 or 10)?
> > Would it be possible on openROADM side to create a revision release (7.2) to bring this kind of minor corrections?
> > I would much prefer corrections on official models comes from openROADM rather than having to tinker with the project models.
> 
> I agree with Gilles that this should better be addressed by OpenROADM in a minor release.
> If we deviate from official models for any reason it should be stated in the models and the commit message.
> 
> JIRA ticket is not enough though it explains the reason.
> Also it is not clear about who made the decision to deviate from official models.

Thanks for your comments. As of now the all the vendors supporting B100G NEs are providing the equipment for 7.1 models. However in 7.1 models there is no regen identity supported (MSA missed to add it). So, instead of creating a minor release just for this, it has been agreed that the vendors and the respective controller teams would add this identity manually, without any new release. Yes, it does deviate from the MSA models, but this deviation has been agreed by the equipment providers. We already started to get equipment that supports regen capabilities for B100G, under 7.1 models. Without this change we will not be able to use T-PCE for 7.1 models.

Also note regen support has been added in the later release (8.1). Yes, we could move to 8.1, but the devices will still be 7.1 and moving T-PCE to 8.1 will not help either.

Since JIRA is a public information, I didn't not put details on the who made the decision and information about the OEMs providing equipment.

Hope this clarifies some of the questions, if not we can discuss during the call today.

Patch-set: 2