Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> Not in favor to disaligne openroadm models of the project from official ones. Would not it be possible to adopt another work around to wait for the next model upgrade (8 or 10)?
> Would it be possible on openROADM side to create a revision release (7.2) to bring this kind of minor corrections?
> I would much prefer corrections on official models comes from openROADM rather than having to tinker with the project models.

Yes, I would prefer as well Gilles, but unfortunately there is no minor or bug release for 7.1. All the equipment provided is on 7.1 models. We can live without this change, but since models would be deviating anyway if regen support was added, I included this change here as well.

Patch-set: 1