Update patch set 3
Patch Set 3: Code-Review+1

Just a general comment, what is good practice in the use of intermediate variables?
In change 97580, you specify several times that it is not necessary to go through an intermediate variable, here it seems not to be the case?

Patch-set: 3
Reviewer: Gerrit User 7085 <7085@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1