Add new method in PortMapping interface
- change the name of deleteMappingData method to deletePortMappingNode
to be more explicit about the deletion of the all port-mapping node when
disconnecting a node
- create a new deleteMapping method that remove a given mapping from the
mapping list for a given node. Such method is necessary when we want to
update a mapping removing one leaf (typically, supporting odu and otu
interface of xponder when we delete ODU4 or OTU4 services).

JIRA: TRNSPRTPCE-538
Signed-off-by: Gilles Thouenon <gilles.thouenon@orange.com>
Change-Id: I477b7a6ddbf3a97b65e12ec317271d45a3a83da3