Modify OpenRoadmInterfaces implementations
When we create a new interface on a device, we need to just after update
the mapping to take into account this new interface. Unfortunatelly, it
may happen sometimes that the new interface is not present yet in the
operational device datastore at the moment of the update mapping occurs.
To avoid such roll-back root cause error, this change modify the
postInterface method to introduce a check of the new interface presence
before returning. As for the deleteInterface method, it also invoques
the update mapping to remove supporting-interfaces.

JIRA: TRNSPRTPCE-539
Signed-off-by: Gilles Thouenon <gilles.thouenon@orange.com>
Change-Id: I61dfc1021a32626341f69922749bf1f9ba9a1062