Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> Please let me know if know the documentation is good or not. Also, if the warning keeps appearing could you help me?? because I have check some online rst editors and there is no error on the text... and the json code blocks are the same as the ones that were previously in the guide... Thanks 😊


Thanks Javier for your effort.

I can give you a hand on fixing gate issues.

But since this documentation is on top of your relation chain, wheteher it passes the gate or not, I won't be able to merge it before the other changes have been approved and for that, I need the other committers'feedback since I am not a TAPI specialist.

I propose you split this change to outsource the part of the documentation that described only what was integrated in Phosphorus and put it at the bottom of the relation chain stack.

Hope this helps
Guillaume

Patch-set: 7
Reviewer: Gerrit User 6731 <6731@7fc14799-209e-464c-9743-7a06c2c21a81>