Update patch set 3
Patch Set 3:

As long as this does not break anything which I do not think it does, it is fine to add it to stable branches. Ultimately it is project decision to include new features/improvements into stable branches.

Patch-set: 3
Reviewer: Gerrit User 797 <797@7fc14799-209e-464c-9743-7a06c2c21a81>