Fix func. tests utils BASE_URL generation issue
RESTCONF base URL is generated by concatenating a string to a
configurable port number.
This port number is loosely typed with an integer.
If this is working well on most systems, it has been reported this
concatenation raises an error on some installation.

Casting explicitly the port number to a string in the concatenation
should solve the issue.

Signed-off-by: guillaume.lambert <guillaume.lambert@orange.com>
Change-Id: Iff503422cf3a881b178618a439d810c918b52953