Keep AvailFreqMaps when netconf session restored
When the netconf session with a ROADM went down and then it came up
again, the node was re-created in the openroadm topology, resetting the
available frequencies bitmaps for both degrees and SRGs. This meannt
that the PCE could allocate already used spectrum for new services,
disrupting existing services.

When creating a new node, first check if it already exists in the
portmapping, and if so, don't initialize available frequencies.

JIRA: TRNSPRTPCE-594

Signed-off-by: Jonas Mårtensson <jonas.martensson@ri.se>
Change-Id: I62f581ed4247b594b14f7c3769e4053b5f23116b