Update Java version check in installMavenCentOS.sh
Java version strings are of the form XX.YY.ZZ
For Java 8 and earlier, XX was always 1 and YY was major version number.
For Java 9 and later, XX is the major version number

OpenDaylight versions that supported Java 8 are now all EOL
and there is no more reason to check YY content.
- update installMavenCentOS.sh to retrieve only the XX parameter
- update shebang to /bin/sh since we are in the area

JIRA: TRNSPRTPCE-700
Change-Id: Id0cff27763d48be773286f1c1a8312bb50e52523
Signed-off-by: guillaume.lambert <guillaume.lambert@orange.com>