Fix persistence when karaf is restarted
OpenDaylight supports datastore persistence but this does not currently
work for TransportPCE since the service list and all openroadm network
layers are initialised as empty when karaf is started.

This changes the transaction from put to merge in order to not
overwrite existing services and network layers.

This also sets config datastore persistence to false before running
tests since they are implemented with the assumption that datastore is
not persistent.

JIRA: TRNSPRTPCE-706

Signed-off-by: Jonas Mårtensson <jonas.martensson@smartoptics.com>
Change-Id: Icec7a1ae47d7001eb408b3b573a837daa8645bc8