Refactor PCE graph PostAlgoPathValidator step 3
checkOSNR() method is very long and has a hich cyclomatic complexity.
The main reason is that it proposes a common heuristic for both AZ & ZA
directions to compute the OSNR. This creates a lot of checks in this
function to identify the direction currently used.
Those multiple checks also affects the code overall efficiency.
- dispatch checkOSNR() content in two methods checkOSNRaz()
  and checkOSNRaz() to check the current direction only once.

The second reason is that it contains a for-loop where the iterator
value is hidden to decorrelate the indexes of the verticles retrieved
by the treatment
and also to back-pedal in the ZA direction case, what is now useless.
- unhide pathElement and merge it to the n iterator
- use decrement operator in the ZA direction for-loop case

Also the first two iterations and the last iteration required
specific treatments since they are at the edges of the line, what
required checks at each iteration.
These other checks affect the code overall efficiency too.
- split the for-loop in three parts to avoid checking iterator value
  at each iteration

The third reason is that the for-loop use a switch case to identify
each node type. Each block contains a different treatment to compute
the signal impairments & characteristics.
- create various intermediate functions to compute signal
  characteristics at each step and reduce case-blocks size.

JIRA: TRNSPRTPCE-725
Signed-off-by: guillaume.lambert <guillaume.lambert@orange.com>
Change-Id: I9cd2370d42d74892f8788a9991827066c2b4fcaa